Removed custom AssetManager and started working on moving everything over to the LibGDX provided AssetManager.
It does not currently use the final planned system, but the base is there to test to ensure it functions.

A few classes need to have a new reset function implemented so that on a new game starting classes, such as the CustomerController, can function correctly as if they had just began functioning. In addition, for the classes that need to be refactored, such as the Recipes and Ingredients, they need to have it before the AssetManager can be completely integrated.

As a result, the program will crash upon trying to access the food Textures that don't yet load in.